backend/flask-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pymongo==4.9.2
motor==3.3.2
pydantic==2.6.4
gunicorn==23.0.0

